# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI Auth App",
    "description": "Ello Bot Backend API",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health Check",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "User Registration",
        "description": "Register a new user with username and password.",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_AuthResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "User Login",
        "description": "Login with username and password, returns user info and access token.",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_AuthResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get Current User",
        "description": "Get the currently authenticated user's information.",
        "operationId": "get_current_user_info_api_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_UserInfoResponse_"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{user_id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get User Info",
        "description": "Get user information by user ID.",
        "operationId": "get_user_info_api_users__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_UserInfoResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/reset": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Reset Password",
        "description": "Reset the password for the currently authenticated user.",
        "operationId": "reset_password_api_users_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_NoneType_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/avatar": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Upload Avatar",
        "description": "Upload a new avatar for the currently authenticated user.",
        "operationId": "upload_avatar_api_users_avatar_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadAvatarRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_NoneType_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuthResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserInfoResponse",
            "description": "User information"
          },
          "token": {
            "type": "string",
            "title": "Token",
            "description": "JWT access token"
          }
        },
        "type": "object",
        "required": [
          "user",
          "token"
        ],
        "title": "AuthResponse",
        "description": "Authentication response schema\n\nContains user information and token returned after successful login or registration.\n\nAttributes:\n    user: User information (sanitized)\n    token: JWT access token"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 20,
            "minLength": 3,
            "title": "Username",
            "description": "Username",
            "examples": [
              "john_doe"
            ]
          },
          "password": {
            "type": "string",
            "maxLength": 20,
            "minLength": 6,
            "title": "Password",
            "description": "Password",
            "examples": [
              "password123"
            ]
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
      "RegisterRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 20,
            "minLength": 3,
            "title": "Username",
            "description": "Username",
            "examples": [
              "john_doe"
            ]
          },
          "password": {
            "type": "string",
            "maxLength": 20,
            "minLength": 6,
            "title": "Password",
            "description": "Password",
            "examples": [
              "password123"
            ]
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "RegisterRequest"
      },
      "ResetPasswordRequest": {
        "properties": {
          "new_password": {
            "type": "string",
            "maxLength": 20,
            "minLength": 6,
            "title": "New Password",
            "description": "New password",
            "examples": [
              "newpassword123"
            ]
          }
        },
        "type": "object",
        "required": [
          "new_password"
        ],
        "title": "ResetPasswordRequest"
      },
      "Result": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "0"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Success"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          }
        },
        "type": "object",
        "title": "Result",
        "description": "Result wrapper for service layer operations\n\nEncapsulates success/failure of an operation along with data or error information.\n\nAttributes:\n    success: Indicates if the operation was successful\n    data: The result data if successful, otherwise None\n    code: Error code if failed, \"0\" if successful\n    message: Error message if failed, \"Success\" if successful"
      },
      "Result_AuthResponse_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "0"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          }
        },
        "type": "object",
        "title": "Result[AuthResponse]"
      },
      "Result_NoneType_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "0"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Success"
          },
          "data": {
            "type": "null",
            "title": "Data"
          },
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          }
        },
        "type": "object",
        "title": "Result[NoneType]"
      },
      "Result_UserInfoResponse_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "0"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserInfoResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          }
        },
        "type": "object",
        "title": "Result[UserInfoResponse]"
      },
      "UploadAvatarRequest": {
        "properties": {
          "avatar_url": {
            "type": "string",
            "title": "Avatar Url",
            "description": "Avatar URL",
            "examples": [
              "https://example.com/avatar.jpg"
            ]
          }
        },
        "type": "object",
        "required": [
          "avatar_url"
        ],
        "title": "UploadAvatarRequest"
      },
      "UserInfoResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "User ID"
          },
          "username": {
            "type": "string",
            "maxLength": 20,
            "minLength": 3,
            "title": "Username",
            "description": "Username",
            "examples": [
              "john_doe"
            ]
          },
          "avatar": {
            "type": "string",
            "title": "Avatar",
            "description": "Avatar URL",
            "default": ""
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "User role",
            "default": "user"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Is active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "username"
        ],
        "title": "UserInfoResponse",
        "description": "User Info response schema (sanitized)\n\nUser information returned to client, does not contain sensitive information.\n\nAttributes:\n    id: User ID\n    username: Username\n    avatar: Avatar URL\n    role: User role\n    is_active: Is active"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}